[Front] 유저 삭제 기능 & 관리자는 삭제 불가능 하게 백앤드 짜놓음

+++ src/main/java/com/FINAL/KIP/user/service/UserService.java
@@ -138,6 +138,8 @@ public class UserService {
     @Transactional
     @UserAdmin
     public void delete(String employeeId){
+        if(employeeId.equals("k-1234567890"))
+            throw new IllegalArgumentException("관리자 계정은 삭제 불가합니다.");
         User userInfo = getUserByEmployeeId(employeeId);
         userRepo.delete(userInfo);
     }
